Created basic Stop model to link Stations and Lines.
So that Lines can be associated with Stations, I created a Stops model which
serves as the :through in their two-way has_many relationship. The Lines
a Station serves can be accessed via @station.lines_served; the Stations
a Line stops at can be accessed via @line.stations. There are a few more
convenient Stop-related functions in the Station and Line classes which I
won't discuss here.

When the transit network graph is implemented, the Stop model may also serve
as the graph's Node model. Transfers might then be represented by specially-marked
edges between Stops which are within the same station. I'm still thinking of
what the best solution to implementing the transit network graph may be.