Improved SearchController behavior.
Search requests with an invalid station name are caught now and an error
message is rendered from the flash. The implementation is pretty ugly, but
I promise myself to go back and clean it up in the near future!

Also, the search#search action was removed and replaced with search#index.
Moreover, if search#index is requested with invalid or no parameters, it
now renders a simple search page.